chore: move stuff from chain-gateway to near-contract-transport
