+++ crates/chain-gateway/Cargo.toml
@@ -6,10 +6,9 @@ edition = { workspace = true }
 autotests = false
 
 [features]
-test-utils = []
+test-utils = ["near-contract-transport/test-utils"]
 
 [dependencies]
-borsh = { workspace = true }
 derive_more = { workspace = true }
 ed25519-dalek = { workspace = true }
 near-account-id = { workspace = true }
@@ -23,11 +22,8 @@ near-mpc-bounded-collections = { workspace = true }
 near-o11y = { workspace = true }
 nearcore = { workspace = true }
 prometheus = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
-tokio-util = { workspace = true }
 tracing = { workspace = true }
 
 [dev-dependencies]
@@ -40,6 +36,7 @@ near-sdk = { workspace = true }
 near-store = { workspace = true }
 rand = { workspace = true }
 rstest = { workspace = true }
+serde_json = { workspace = true }
 tempfile = { workspace = true }
 test-port-allocator = { workspace = true }
 test-utils = { workspace = true }
