+++ crates/chain-gateway/src/lib.rs
@@ -4,7 +4,6 @@ pub mod event_subscriber;
 #[cfg(any(test, feature = "test-utils"))]
 pub mod mock;
 pub mod primitives;
-pub mod state_viewer;
 pub mod transaction_sender;
 pub mod types;
