--- crates/chain-gateway/src/state_viewer.rs
@@ -1,7 +1,0 @@
-mod monitoring;
-mod subscription;
-mod traits;
-mod view_call;
-
-pub use traits::ViewExt;
-pub use view_call::ViewCall;
