+++ crates/near-contract-transport/Cargo.toml
@@ -7,7 +7,16 @@ repository.workspace = true
 description = "NEAR function-call payload and the CallContract/ViewContract transport traits shared across the MPC stack"
 
 [features]
-traits = ["dep:near-account-id", "dep:borsh", "dep:thiserror", "dep:serde_json"]
+traits = [
+    "dep:near-account-id",
+    "dep:borsh",
+    "dep:thiserror",
+    "dep:serde_json",
+    "dep:tokio",
+    "dep:tokio-util",
+    "dep:tracing",
+]
+test-utils = ["traits"]
 
 [dependencies]
 borsh = { workspace = true, optional = true }
@@ -18,6 +27,13 @@ near-token = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true, optional = true }
 thiserror = { workspace = true, optional = true }
+tokio = { workspace = true, optional = true }
+tokio-util = { workspace = true, optional = true }
+tracing = { workspace = true, optional = true }
+
+[dev-dependencies]
+assert_matches = { workspace = true }
+rstest = { workspace = true }
 
 [lints]
 workspace = true
