+++ crates/near-contract-transport/src/views.rs
@@ -1,11 +1,15 @@
 mod args;
 mod deserialize;
 mod errors;
+mod monitoring;
 mod observation;
+mod subscription;
 mod traits;
+mod view_call;
 
 pub use args::ViewArgs;
-pub use deserialize::{DeserializationError, Deserializer, borsh_de, json_de};
+pub use deserialize::{DeserializationError, Deserializer};
 pub use errors::TransportError;
 pub use observation::{ObservedState, SerializedObservation};
 pub use traits::{HasPollInterval, PollInterval, ViewContract, WatchContractState};
+pub use view_call::ViewCall;
