+++ crates/tee-context/Cargo.toml
@@ -21,6 +21,7 @@ assert_matches = { workspace = true }
 chain-gateway = { workspace = true, features = ["test-utils"] }
 ed25519-dalek = { workspace = true }
 mpc-primitives = { workspace = true }
+near-contract-transport = { workspace = true, features = ["test-utils"] }
 serde_json = { workspace = true }
 tokio = { workspace = true, features = ["test-util", "macros"] }
