+++ docs/tee-lifecycle.md
@@ -23,15 +23,15 @@ Services run inside Dstack CVMs, booted through a [Launcher][launcher] that meas
 | [Transaction Sender][transaction-sender] | Submits attestation transactions to the governance contract |
 
 [dstack]: https://github.com/Dstack-TEE/dstack
-[chain-indexer]: chain-gateway-design.md
+[chain-indexer]: design/chain-gateway-design.md
 [tee-authority]: https://github.com/near/mpc/tree/ce53324f472aa89fdf702d7482211bbdb6a44967/crates/tee-authority
 [`tee-authority`]: https://github.com/near/mpc/tree/ce53324f472aa89fdf702d7482211bbdb6a44967/crates/tee-authority
 [mpc-attestation]: https://github.com/near/mpc/blob/ce53324f472aa89fdf702d7482211bbdb6a44967/crates/mpc-attestation/src/attestation.rs#L29
 [`mpc-attestation`]: https://github.com/near/mpc/blob/ce53324f472aa89fdf702d7482211bbdb6a44967/crates/mpc-attestation/src/attestation.rs#L29
-[contract-state-subscriber]: chain-gateway-design.md#state-viewer
-[transaction-sender]: chain-gateway-design.md#transaction-sender
+[contract-state-subscriber]: design/chain-gateway-design.md#state-viewer
+[transaction-sender]: design/chain-gateway-design.md#transaction-sender
 [tee-context-design]: tee-context-design.md
-[mpc-context]: chain-gateway-design.md
+[mpc-context]: design/chain-gateway-design.md
 [launcher]: securing-mpc-with-tee-design-doc.md#launcher-pattern
 
 ### Crate Dependencies
@@ -142,7 +142,7 @@ Individual services may add steps between "Start application container" and the 
 
 The [TEE Context][tee-context-design] is a shared crate managing the TEE attestation lifecycle. The MPC node already implements the attestation tasks in [`remote_attestation.rs`][remote-attestation] and [`allowed_image_hashes_watcher.rs`][allowed-hashes-watcher]; they will be extracted into a standalone crate, depending on [`tee-authority`][tee-authority] and [`mpc-attestation`][mpc-attestation], reusable by all services.
 
-[mpc-context]: chain-gateway-design.md
+[mpc-context]: design/chain-gateway-design.md
 
 [remote-attestation]: https://github.com/near/mpc/blob/ce53324f472aa89fdf702d7482211bbdb6a44967/crates/node/src/tee/remote_attestation.rs
 [allowed-hashes-watcher]: https://github.com/near/mpc/blob/ce53324f472aa89fdf702d7482211bbdb6a44967/crates/node/src/tee/allowed_image_hashes_watcher.rs#L103
